Move recursive cache update to apollo/update
